OpenCV(YOLO)/requirements.txt
torch
torchvision
numpy
pillow
matplotlib
tqdm
fastapi
uvicorn[standard]
httpx
opencv-python
requests
xgboost>=1.7.0
pandas>=1.5.0
scikit-learn>=1.2.0
googlemaps>=4.10.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
